build(demos/astro-js-sass-styling): 🧱 update dependencies
┌────────────────────────┬─────────┬─────────┬───────────────────────┐
│ Package                │ From    │ To      │ Dependents            │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ postcss (dev)          │ 8.4.47  │ 8.4.49  │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ prettier (dev)         │ 3.3.3   │ 3.4.2   │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ sass (dev)             │ 1.80.6  │ 1.83.0  │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ sass-migrator          │ 2.2.1   │ 2.3.0   │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ stylelint (dev)        │ 16.10.0 │ 16.12.0 │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ svelte (dev)           │ 5.7.1   │ 5.15.0  │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ typescript (dev)       │ 5.6.3   │ 5.7.2   │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ @astrojs/react (dev)   │ 3.6.3   │ 4.1.1   │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ @astrojs/svelte (dev)  │ 5.7.3   │ 7.0.1   │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ @types/react (dev)     │ 18.3.12 │ 19.0.2  │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ @types/react-dom (dev) │ 18.3.1  │ 19.0.2  │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ astro (dev)            │ 4.16.10 │ 5.1.0   │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ react (dev)            │ 18.3.1  │ 19.0.0  │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ react-dom (dev)        │ 18.3.1  │ 19.0.0  │ astro-js-sass-styling │
├────────────────────────┼─────────┼─────────┼───────────────────────┤
│ vite (dev)             │ 5.4.11  │ 6.0.4   │ astro-js-sass-styling │
└────────────────────────┴─────────┴─────────┴───────────────────────┘
