--- demos/astro-js-sass-styling/.prettierrc
@@ -1,7 +1,0 @@
-{
-	"useTabs": true,
-	"arrowParens": "always",
-	"singleQuote": true,
-	"trailingComma": "all",
-	"printWidth": 100
-}
