+++ demos/astro-js-sass-styling/.prettierrc.mjs
@@ -1,0 +1,8 @@
+/** @type {import("prettier").Config} */
+export default {
+  plugins: ["prettier-plugin-astro", "prettier-plugin-svelte"],
+  overrides: [
+    { files: "*.astro", options: { parser: "astro" } },
+    { files: "*.svelte", options: { parser: "svelte" } },
+  ],
+};
