+++ demos/astro-js-sass-styling/.stylelintrc.json
@@ -1,25 +1,28 @@
 {
-	"extends": "stylelint-config-recommended-scss",
-	"rules": {
-		"color-named": "never",
-		"font-family-name-quotes": "always-where-required",
-		"font-weight-notation": "named-where-possible",
-		"function-url-no-scheme-relative": true,
-		"function-url-quotes": "always",
-		"value-keyword-case": "lower",
-		"unit-disallowed-list": [],
-		"no-descending-specificity": true,
-		"no-duplicate-selectors": true,
-		"font-family-no-missing-generic-family-keyword": null,
-		"property-no-unknown": [
-			true,
-			{
-				"ignoreProperties": ["/^lost-/"]
-			}
-		],
-		"selector-pseudo-class-no-unknown": [true, { "ignorePseudoClasses": ["global"] }]
-	},
-	"ignoreFiles": ["node_modules/*"],
-	"defaultSeverity": "error",
-	"customSyntax": "postcss-html"
+  "extends": "stylelint-config-recommended-scss",
+  "rules": {
+    "color-named": "never",
+    "font-family-name-quotes": "always-where-required",
+    "font-weight-notation": "named-where-possible",
+    "function-url-no-scheme-relative": true,
+    "function-url-quotes": "always",
+    "value-keyword-case": "lower",
+    "unit-disallowed-list": [],
+    "no-descending-specificity": true,
+    "no-duplicate-selectors": true,
+    "font-family-no-missing-generic-family-keyword": null,
+    "property-no-unknown": [
+      true,
+      {
+        "ignoreProperties": ["/^lost-/"]
+      }
+    ],
+    "selector-pseudo-class-no-unknown": [
+      true,
+      { "ignorePseudoClasses": ["global"] }
+    ]
+  },
+  "ignoreFiles": ["node_modules/*"],
+  "defaultSeverity": "error",
+  "customSyntax": "postcss-html"
 }
