+++ demos/astro-js-sass-styling/.vscode/extensions.json
@@ -1,4 +1,4 @@
 {
-	"recommendations": ["astro-build.astro-vscode"],
-	"unwantedRecommendations": []
+  "recommendations": ["astro-build.astro-vscode"],
+  "unwantedRecommendations": []
 }
