+++ demos/astro-js-sass-styling/astro.config.mjs
@@ -1,9 +1,22 @@
-import { defineConfig } from 'astro/config';
+import { defineConfig } from "astro/config";
 
-import react from '@astrojs/react';
-import svelte from '@astrojs/svelte';
+import react from "@astrojs/react";
+import svelte from "@astrojs/svelte";
 
 // https://astro.build/config
 export default defineConfig({
-	integrations: [react(), svelte()],
+  vite: {
+    css: {
+      preprocessorOptions: {
+        scss: {
+          additionalData: "@use 'variables.scss';",
+          loadPaths: ["src/styles"],
+        },
+      },
+    },
+  },
+  integrations: [
+    react({ include: ["**/react/*"] }),
+    svelte({ include: ["**/svelte/*"] }),
+  ],
 });
