+++ demos/astro-js-sass-styling/src/styles/styles.scss
@@ -1,67 +1,67 @@
 @charset "UTF-8";
 
-@use 'fonts.scss';
-@use 'variables.scss';
+@use "fonts.scss";
 
 body {
-	margin: variables.$spacing-0;
-	background: hsl(
-		variables.$colour-dark-hue variables.$colour-dark-saturation variables.$colour-dark-luminance
-	);
-	font-family: Roboto;
-	text-align: center;
+  margin: variables.$spacing-0;
+  background: hsl(
+    variables.$colour-dark-hue variables.$colour-dark-saturation
+      variables.$colour-dark-luminance
+  );
+  font-family: Roboto;
+  text-align: center;
 }
 
 h1 {
-	font-size: 3.815rem;
-	margin-bottom: variables.$spacing-12;
+  font-size: 3.815rem;
+  margin-bottom: variables.$spacing-12;
 }
 
 h2 {
-	font-size: 1.802rem;
-	margin-bottom: variables.$spacing-8;
+  font-size: 1.802rem;
+  margin-bottom: variables.$spacing-8;
 }
 
 p {
-	margin: variables.$spacing-6 variables.$spacing-6 variables.$spacing-8;
+  margin: variables.$spacing-6 variables.$spacing-6 variables.$spacing-8;
 }
 
 button {
-	cursor: pointer;
-	margin: variables.$spacing-6;
-	border: solid variables.$spacing-px
-		hsl(
-			variables.$colour-light-text-hue variables.$colour-light-text-saturation
-				variables.$colour-light-text-luminance
-		);
-	border-radius: 50%;
-	padding: variables.$spacing-8;
+  cursor: pointer;
+  margin: variables.$spacing-6;
+  border: solid variables.$spacing-px
+    hsl(
+      variables.$colour-light-text-hue variables.$colour-light-text-saturation
+        variables.$colour-light-text-luminance
+    );
+  border-radius: 50%;
+  padding: variables.$spacing-8;
 }
 
 .video-container {
-	position: relative;
-	aspect-ratio: 16 / 9;
-	width: 100%;
-	max-width: 560px;
-	margin: variables.$spacing-0 variables.$spacing-6;
+  position: relative;
+  aspect-ratio: 16 / 9;
+  width: 100%;
+  max-width: 560px;
+  margin: variables.$spacing-0 variables.$spacing-6;
 }
 
 .video-container iframe {
-	position: absolute;
-	inset: 0;
-	width: 100%;
-	height: 100%;
-	border: 0;
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  border: 0;
 }
 
 .screen-reader-text {
-	border: 0;
-	clip: rect(1px, 1px, 1px, 1px);
-	clip-path: inset(50%);
-	height: 1px;
-	margin: -1px;
-	width: 1px;
-	overflow: hidden;
-	position: absolute !important;
-	word-wrap: normal !important;
+  border: 0;
+  clip: rect(1px, 1px, 1px, 1px);
+  clip-path: inset(50%);
+  height: 1px;
+  margin: -1px;
+  width: 1px;
+  overflow: hidden;
+  position: absolute !important;
+  word-wrap: normal !important;
 }
