+++ demos/astro-js-sass-styling/tsconfig.json
@@ -1,9 +1,9 @@
 {
-	"compilerOptions": {
-		"baseUrl": ".",
-		"paths": {
-			"~*": ["src/*"]
-		}
-	},
-	"extends": "astro/tsconfigs/strict"
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "~*": ["src/*"]
+    }
+  },
+  "extends": "astro/tsconfigs/strict"
 }
